fix: restructure the app
