+++ src/lib/types/linkedinInfo.ts
@@ -1,0 +1,8 @@
+export interface LinkedinUserInfo {
+  country: number;
+  language: number;
+  name: string;
+  email: string;
+  picture: string;
+  email_verified: boolean;
+}
