+++ src/lib/types/optional.ts
@@ -1,0 +1,1 @@
+export type Optional<T> = T | undefined;
